requirements.txt
publicdotcom-py>=0.1.0
httpx>=0.28,<1
PyYAML>=6.0,<7
APScheduler>=3.10,<4
pytest>=8.0,<9
python-dotenv>=1.0,<2
